Define locals.tags variable for aws_instances data filter
To retrieve the public ip addresses from the created wireguards virtual machine
define a locals.tags that is used in the data provider `aws_instances` to find
this instances and extract the public ip address within terraform.

Then the `data.aws_instances.wireguards` provider is used to define the output
variable `wireguards_eip` so that the public elastic ip address can be queried
with `terraform output` command for example.