Simplify the wireguard server configuration setup.
The assignment of the elastic ip address is done by the aws auto scaling group
if the terraform module variable `var.use_eip` is set to true. The elastic ip
address assignment that were part of the user-data script is completely removed.
This make the whole process quite simple and all the iam policy assignment for the
wireguard virtual machine are obsolete and therefore discarded.

The wireguard server address is retrieved within the user-data script so no
need to specify it in terraform and the `wg_server_net` terraform variable
is removed.

The wireguard server private/public key is generated within the user-data script.
That make the `aws_ssm_parameter` declaration obsolete.