Add option to secure ssh access with VPN connection
The terraform variable `vpn_enabled_ssh=true` secure the ssh port with
the network cidr of the wireguard server network range. That means
in order to access the wrieguard server with ssh an wireguard VPN connection
to the same server is needed.

If this variable is set to `vpn_enabled_ssh=false` then the ssh port
of the wireguard server is public available.

The default value is `true`.